pyproject.toml
[project]
name = "08-tmr-update-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "geopy>=2.4.1",
    "groq>=0.30.0",
    "loguru>=0.7.3",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "shapely>=2.1.1",
    "supabase>=2.6.0",
]
